Utils file refactoring
